+++ src/main/java/uk/gov/hmcts/reform/pcs/ccd/repository/enforcetheorder/WritRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import uk.gov.hmcts.reform.pcs.ccd.entity.enforcetheorder.WritEntity;
 
 
-public interface WritRepository extends JpaRepository<WritEntity, Long> {
+public interface WritRepository extends JpaRepository<WritEntity, Integer> {
 }
