+++ src/test/java/uk/gov/hmcts/reform/pcs/ccd/service/defenceform/DefenceFormSchedulerTest.java
@@ -21,7 +21,7 @@ import static org.mockito.Mockito.when;
 class DefenceFormSchedulerTest {
 
     private static final long CASE_REFERENCE = 1234567812345678L;
-    private static final Long RESPONSE_ID = 1L;
+    private static final Integer RESPONSE_ID = 1;
     private static final UUID PARTY_ID = UUID.randomUUID();
 
     @Mock
