+++ src/test/java/uk/gov/hmcts/reform/pcs/ccd/service/enforcetheorder/strategy/WarrantStrategyTest.java
@@ -96,7 +96,7 @@ class WarrantStrategyTest {
         // Given
         WarrantEntity warrantEntity = new WarrantEntity();
         WarrantEntity savedWarrantEntity = new WarrantEntity();
-        savedWarrantEntity.setId(1L);
+        savedWarrantEntity.setId(1);
 
         when(warrantDetailsMapper.toEntity(enforcementOrder, enforcementOrderEntity))
             .thenReturn(warrantEntity);
@@ -118,7 +118,7 @@ class WarrantStrategyTest {
         // Given
         WarrantEntity warrantEntity = new WarrantEntity();
         WarrantEntity savedWarrantEntity = new WarrantEntity();
-        Long warrantId = 1L;
+        Integer warrantId = 1;
         savedWarrantEntity.setId(warrantId);
 
         when(warrantDetailsMapper.toEntity(enforcementOrder, enforcementOrderEntity))
